new siridium character starbound
we have new plans for a new ship, as well as better managing our character. We have our items from our old character coming along to our new one soon. This will be a more legit gameplay. Note: some of our items are  next to the mine on the current world.